requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Bootstrap==3.3.7.1
twilio==8.5.0
python-dotenv==1.0.0
email-validator==2.1.0
werkzeug==2.3.7
gunicorn==21.2.0
